feat: pr 피드백 적용 및 가입되어 있지 않은 유저가 channel 목록 보지 못하게 변경

+++ discodeit/src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -116,15 +116,10 @@ public class BasicMessageService implements MessageService {
   @Transactional
   @Override
   public void delete(UUID messageId) {
-    Message message = messageRepository.findById(messageId)
-        .orElseThrow(
-            () -> new NoSuchElementException("Message with id " + messageId + " not found"));
-
-    List<UUID> attachmentIds = message.getAttachments().stream().map(BinaryContent::getId).toList();
-    messageRepository.deleteById(messageId);
-    if (!attachmentIds.isEmpty()) {
-      binaryContentRepository.deleteAllByIdInBatch(attachmentIds);
+    if (!messageRepository.existsById(messageId)) {
+      throw new NoSuchElementException("Message with id " + messageId + " not found");
     }
+    messageRepository.deleteById(messageId);
   }
 
 }
